# Implemented PoolOptimizer