--- a/src/Composer/DependencyResolver/Pool.php
+++ b/src/Composer/DependencyResolver/Pool.php
@@ -221,6 +221,14 @@ class Pool implements \Countable
         return \in_array($package, $this->unacceptableFixedOrLockedPackages, true);
     }
 
+    /**
+     * @return BasePackage[]
+     */
+    public function getUnacceptableFixedOrLockedPackages()
+    {
+        return $this->unacceptableFixedOrLockedPackages;
+    }
+
     public function __toString()
     {
         $str = "Pool:\n";