--- a/tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/multi-repo-replace-partial-update.test
+++ b/tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/multi-repo-replace-partial-update.test
@@ -105,3 +105,13 @@ Check that replacers from additional repositories are loaded when doing a partia
     "shared/dep-1.0.0.0",
     "shared/dep-1.2.0.0"
 ]
+
+--EXPECT-OPTIMIZED--
+[
+    "indirect/replacer-1.2.0.0",
+    "indirect/replacer-1.0.0.0",
+    "replacer/package-1.2.0.0",
+    "replacer/package-1.0.0.0",
+    "base/package-1.0.0.0",
+    "shared/dep-1.2.0.0"
+]