… .github/workflows/continuous-integration.yml
… doc/articles/troubleshooting.md
… src/Composer/DependencyResolver/Problem.php
… src/Composer/DependencyResolver/Rule.php
… src/Composer/Repository/RepositorySet.php
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/partial-update-unfixing-locked-deps.test
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/partial-update-unfixing-with-replacers.test
… tests/Composer/Test/DependencyResolver/Fixtures/pooloptimizer/aliases.test
… tests/Composer/Test/DependencyResolver/Fixtures/pooloptimizer/basic-prefer-highest.test
… tests/Composer/Test/DependencyResolver/Fixtures/pooloptimizer/basic-prefer-lowest.test
… tests/Composer/Test/DependencyResolver/Fixtures/pooloptimizer/conflict.test
… tests/Composer/Test/DependencyResolver/Fixtures/pooloptimizer/conflict2.test
… tests/Composer/Test/DependencyResolver/Fixtures/pooloptimizer/group-by-required.test
… tests/Composer/Test/DependencyResolver/Fixtures/pooloptimizer/locked-fixed-untouched.test
… tests/Composer/Test/DependencyResolver/Fixtures/pooloptimizer/replaces.test
… tests/Composer/Test/DependencyResolver/PoolBuilderTest.php
… tests/Composer/Test/DependencyResolver/PoolOptimizerTest.php
… tests/Composer/Test/Fixtures/installer/conflict-downgrade-nested.test
… tests/Composer/Test/Fixtures/installer/conflict-downgrade.test
… tests/Composer/Test/Fixtures/installer/deduplicate-solver-problems.test
… tests/Composer/Test/Fixtures/installer/github-issues-7051.test
… tests/Composer/Test/Fixtures/installer/provider-conflicts3.test
… tests/Composer/Test/InstallerTest.php
… tests/bootstrap.php
… tests/deprecations-8.1.json