This PR optimizes the `Pool` before the rules are created and passed on to the `Solver`. Every package added to the pool results in an exponential increase of the number of rules the `Solver` has to analyze so the best we can do is to try to throw away all the packages that aren't of any interest early on.

I've implemented a new `PoolOptimizer` and set up a test framework to which we can add more optimizations in the future but the current implementation already provides **huge** benefits!

It works by early selecting the correct package based on the provided policy. It's common to have loads of bugfix releases that all have the same dependency definition. If you imagine e.g. `symfony/routing` in its `4.4` LTS branch, all of them hardly ever define new `require`, `replace`, `provide` or `conflict` definitions. Chances that `4.4.0` and `4.4.14` are the same, are really high and so we don't need to keep all of the package versions in between, we can drop the irrelevant ones depending on `--prefer-stable` and `--prefer-lowest`. There's a bit more to it and it's not just that simple but that's the basic idea behind the first optimization in the new `PoolOptimizer`.

And here are the **very promising** results:

Contao Managed Edition (https://github.com/contao/managed-edition/blob/master/composer.json):

|  | 1.10.13 | master@5df1797d2 | this PR |
|------|----------:|-----------------:|--------:|
| RAM | 3351.08MiB | 1518.31MiB | 134.54MiB |
| Time | 84s | 28.95s | 3.25s |

Magento (https://github.com/magento/magento2/blob/2.4-develop/composer.json):

|  | 1.10.13 | master@5df1797d2 | this PR |
|------|----------:|-----------------:|--------:|
| RAM | 961.86MiB | 169.64MiB | 46.39MiB |
| Time | 24.61s | 4.38s | 1.17s |

Laravel 7.x (https://github.com/laravel/laravel/blob/7.x/composer.json):

|  | 1.10.13 | master@5df1797d2 | this PR |
|------|----------:|-----------------:|--------:|
| RAM | 911.97MiB | 152.42MiB | 37.62MiB |
| Time | 8.81s | 1.51s | 0.95s |

Packagist (https://github.com/composer/packagist/blob/master/composer.json)

|  | 1.10.13 | master@5df1797d2 | this PR |
|------|----------:|-----------------:|--------:|
| RAM | 3308.18MiB | 1503.63MiB | 134.73M |
| Time | 51.63s | 14.9s | 4.04s |

Of course, the more complex the dependencies (or just the more dependencies in general) the higher the benefits of this PR.

I was too lazy to calculate the benefits in % but I guess we can summarize it like so: 🚀 

I'm off for a few days now so I might not be replying to questions quickly but would love to see others giving it a try and share some numbers :) 

For the lazy folks: Here's a precompiled phar (c996457a039b2f0a5eee3b6364dfce307c614737):

[composer.phar.zip](https://github.com/composer/composer/files/5484474/composer.phar.zip)