@stof 2020-10-02
regarding the time where the pool optimizer is called, I'm not sure it is the best one compared to the `PRE_POOL_CREATE` event triggered for plugins. If plugins are modifying the pool packages, that might break the invariants of the optimization if it is already done.
I think it would be safer to optimize the pool just before entering the solver, but after plugins have tweaked it.