@stof 2020-10-02
ah sorry, I missed that part about the hash.