@hussainweb 2020-10-03
I just wanted to share results of my tests with Drupal. Before each test, I removed the `vendor`, `web/core` (Drupal core) directories, and the `composer.lock` file.

A not-very-complex Drupal 9.1 dev project ([this repo](https://github.com/hussainweb/drupal-php8-composer2))

-|Composer 1.10.3 | Composer 2.0 RC1 | This PR (phar file)
-|----------------|------------------|--------------------
Memory|482.7MiB (peak: 2015.39MiB)|60.74MiB (peak: 699.45MiB)|34.28MiB (peak: 60.42MiB)
Time|136.12s|36.06s|31.43s

[Drupal 9.1 core (dev release)](https://git.drupalcode.org/project/drupal)

-|Composer 1.10.3 | Composer 2.0 RC1 | This PR (phar file)
-|----------------|------------------|--------------------
Memory|419.74MiB (peak: 1005.93MiB)|31.03MiB (peak: 214.35MiB)|28MiB (peak: 38.14MiB)
Time|24.77s|13.06s|10.61s

I can see the time improvement is not significant but memory improvement is. This is awesome work. Thanks! 👏