@hussainweb 2020-10-03
I also tried this on a project with Drupal 8; this time running `composer update --profile`. I ran it once before the actual tests so that caches get created, etc. This project had many out-of-date dependencies so there were many packages that got updated. I didn't clear out `vendor` directory for each run but before each run, I restored `composer.lock` to the previous state.

-|Composer 1.10.13 | Composer 2.0 RC1 | This PR (phar file)
-|----------------|------------------|--------------------
Memory|625.66MiB (peak: 2234.86MiB)|133.01MiB (peak: 785.85MiB)|112.02MiB (peak: 158.98MiB)
Time|43.57s|23.27s|21.38s

Again, not a significant differnce in time (this could be because it actually times updating packages on disk) but a significant reduction in peak memory.