@wouterj 2020-10-04
A Symfony application using the Symfony Flex plugin (needed because our Symfony packages use `*` as version constraint):

| - | Composer 1.10.13 | Composer 2.0 RC1 | This PR (phar file) |
| --- | --- | --- | --- |
| Memory | 303.22MiB (peak: 859.14MiB) | 32.69MiB (peak: 79.3MiB) | 32.92MiB (peak: 80.23MiB) |
| Time |12.62s | 8.29s | time: 8.45s |

Not much changes, which I think it expected as the Symfony Flex plugin already significantly reduces the versions that are considered.

---

A Neos base distribution - the most complex project on my PC that isn't requiring Symfony Flex:

| - | Composer 1.10.13 | Composer 2.0 RC1 | This PR (phar file) |
| --- | --- | --- | --- |
| Memory | 413.34MiB (peak: 1321.35MiB) | 46.11MiB (peak: 275.17MiB) | 25.32MiB (peak: 98.34MiB) |
| Time | 20.56s | 11.86s | 9.24s |

Similar results as Drupal 8, not a significant step in performance but a significant reduction of peak memory usage.