@simonberger 2020-10-06
Good job man ! I tested this two days ago with almost same runtime. Now this PR is a lot faster. No idea why :)
Another thing is the memory usage. Really impressive. Improvements in this area in my opinion not less important than the runtime.

|| master@5df1797d2 | this PR
 -- | -- | --
RAM | 1376.52MiB | 85.65MiB
Time | 12.47s | 4.4s